Change an automatically generated int field to double
allow fit to the data with decimal values, fix error in parsing

+++ app/src/main/java/com/mariebyleen/weather/current_conditions/model/CurrentConditionsResponseCoord.java
@@ -1,10 +1,10 @@
 package com.mariebyleen.weather.current_conditions.model;
 
 public class CurrentConditionsResponseCoord {
-    private int lon;
-    private int lat;
+    private double lon;
+    private double lat;
 
-    public int getLon() {
+    public double getLon() {
         return this.lon;
     }
 
@@ -12,7 +12,7 @@ public class CurrentConditionsResponseCoord {
         this.lon = lon;
     }
 
-    public int getLat() {
+    public double getLat() {
         return this.lat;
     }
